String publishers / subscriber ported.
